modifAPI

+++ APIFinale/fonctions.php
@@ -77,6 +77,9 @@ function validationmdp($mdp, $mdpConfirmation)
 function verificationPremiereInscription($PersonnelIdentite)
 {
   $BD = connexionBD();
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
   $recherchePresence = $BD->prepare('SELECT * FROM apprenti WHERE nom = ?');
   $recherchePresence->execute(array($PersonnelIdentite['7']));
   $BD = null;
@@ -98,11 +101,11 @@ function envoiMail($adresseMail, $nouveauMotDePasse)
   mail($destinataire, $sujet, $message, $headers);
 }
 
-function recuperation_role($login)
+function recuperation_role($login, $mdp)
 {
   $BD = connexionBD();
-  $recuperationRoleMembre = $BD->prepare('SELECT id_utilisateur FROM utilisateur WHERE login = ?');
-  $recuperationRoleMembre->execute(array($login));
+  $recuperationRoleMembre = $BD->prepare('SELECT id_role FROM role WHERE login = ? and mdp = ?');
+  $recuperationRoleMembre->execute(array($login, $mdp));
   $BD = null;
   if ($recuperationRoleMembre->rowCount() > 0) {
     foreach ($recuperationRoleMembre as $row) {
@@ -129,11 +132,29 @@ function clean($champEntrant)
 function listeApprenti()
 {
   $BD = connexionBD();
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
+  
+  $listeApprenti = $BD->prepare('SELECT * FROM apprentis');
+  $listeApprenti->execute(array());
+    $BD = null;
+    $resultat = [];
+
+    foreach($listeApprenti as $row){
+        array_push($resultat, array('nom' => $row['nom'], 'prenom' => $row['prenom'], 'photo' => $row['photo']));
+    }
+
+    return $resultat;
 }
+
 
 function unApprenti()
 {
   $BD = connexionBD();
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
 }
 
 
@@ -146,6 +167,10 @@ function unApprenti()
 function inscriptionApprenti($apprentiIdentite)
 {
   $BD = connexionBD();
+  
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
   if (count($apprentiIdentite) > 0) {
     $identiteApprentiHTML = conversionHTML($apprentiIdentite);
 
@@ -172,6 +197,9 @@ function inscriptionApprenti($apprentiIdentite)
 
 function supprimerApprenti($id_apprenti) {
   $BD = connexionBD();
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
   $id_personnel = htmlspecialchars($id_apprenti);
   $suppressionApprenti = $BD->prepare('DELETE FROM apprenti WHERE id_apprenti = ?');
   $suppressionApprenti->execute(array($id_personnel));
@@ -185,6 +213,11 @@ function supprimerApprenti($id_apprenti) {
 function inscriptionPersonnel($PersonnelIdentite)
 {
   $BD = connexionBD();
+  
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
+
   if (count($PersonnelIdentite) > 0) {
     $identitePersonnelHTML = conversionHTML($PersonnelIdentite);
     if ($PersonnelIdentite['6'] == 2) {
@@ -217,6 +250,11 @@ function inscriptionPersonnel($PersonnelIdentite)
 function supprimerPersonnel($id_personnel)
 {
   $BD = connexionBD();
+
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
+
   $id_personnel = htmlspecialchars($id_personnel);
   $suppressionPersonnel = $BD->prepare('DELETE FROM personnel WHERE id_personnel = ?');
   $suppressionPersonnel->execute(array($id_personnel));
@@ -231,6 +269,11 @@ function supprimerPersonnel($id_personnel)
 function modifierPersonnel($personnel, $id_personnel, $nom, $prenom)
 {
   $BD = connexionBD();
+  // Vérifier la connexion
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
+
   $id_personnel = htmlspecialchars($id_personnel);
   $nom = htmlspecialchars($nom);
   $prenom = htmlspecialchars($prenom);
@@ -244,9 +287,17 @@ function modifierPersonnel($personnel, $id_personnel, $nom, $prenom)
   }
 }
 
-function listePersonnel()
+function listePersonnel($role)
 {
   $BD = connexionBD();
+
+  if ($BD->connect_error) {
+    die("Échec de la connexion à la base de données : " . $BD->connect_error);
+  }
+  $listePersonnel = $BD->prepare('SELECT FROM * from personnel');
+  $listePersonnel->execute(array());
+  $BD = null;
+
 }
 
 function unPersonnel()
